+++ web-wallet/src/lib/containers/index.js
@@ -1,4 +1,3 @@
 export { default as AllocateContract } from "./AllocateContract/AllocateContract.svelte";
 export { default as MigrateContract } from "./MigrateContract/MigrateContract.svelte";
 export { default as StakeContract } from "./StakeContract/StakeContract.svelte";
-export { default as TransferContract } from "./TransferContract/TransferContract.svelte";